update site footer hover, update site-menu hover
Changed the active class color for the site-menu, and changed the hover effect on links in site menu.

Added a different hover effect to the links in the footer.

Links in footer have a text-decoration on them to show they are different from the site-menu links. They will lead to a page off of the website rather than a new page on the website.