Fix bug relating to undefined localstorage +
clickEvent.target.matches("button”) was the issue. There is another button on the page, the submit button. When a user would click the submit button this clickEvent would fire, and since there was no theme-attribute localstorage would set it as undefined.

Making the clickEvent more specific to only target something with that data-attribute fixes.